enforce max length in reader
AllenNLP does support embedding seqs > 512 which is why we weren't doing this before, but the code we're using to insert a pseudotoken is incompatible with it at the moment. Since long seqs are so rare, we can just cut them down and be confident it won't affect performance too much.